Rewrite Psalm to avoid dynamic features rejected by the TypePHP AOT compiler
Work in progress towards compiling Psalm with swoole/typephp.

- Move the bootstrap require_once statements out of the CLI classes into
  the launcher scripts, so no source file has executable code at file scope.
- Add Psalm\Internal\TypePhp\Dynamic::any(), an identity helper that keeps a
  local dynamic for the compiler while Psalm still sees the exact type, and
  use it for locals that are reassigned with sibling classes or passed by
  reference.
- Declare object by-reference parameters as `mixed &$x` with strict
  @param/@param-out docblocks (TypePHP only supports references to scalars,
  arrays, mixed and unions).
- Rename loop variables that reused a differently typed local, split
  builder/freeze round-trips into separate variables, terminate switch
  cases that ended in an if, and replace intentional fallthrough in the
  path tokenizer and the memory-limit parser with explicit logic.
- Replace `~` on strings with a byte-wise helper, remove the $GLOBALS-based
  include path resolution, avoid list destructuring in foreach targets and
  loop conditions, and hoist typed property reads that the compiler
  mishandles in lowered ternaries.

+++ src/Psalm/CodeLocation.php
@@ -253,7 +253,8 @@ class CodeLocation
 
             $key_line = trim((string) preg_replace('@\**/\s*@', '', mb_strcut($key_line, $indentation)));
 
-            $this->selection_start = $preview_offset + $indentation + $this->preview_start;
+            /** @psalm-suppress RedundantCast the cast keeps the dynamic sum assignable to a typed property under TypePHP */
+            $this->selection_start = (int) ($preview_offset + $indentation + $this->preview_start);
             $this->selection_end = $this->selection_start + strlen($key_line);
         }
 
@@ -287,7 +288,8 @@ class CodeLocation
                         . "or regex doesn't contain 1st capturing group, regex type " . $this->regex_type,
                     );
                 }
-                $this->selection_start = $this->selection_start + $matches[1][1];
+                /** @psalm-suppress RedundantCast the cast keeps the dynamic sum assignable to a typed property under TypePHP */
+                $this->selection_start = (int) ($this->selection_start + $matches[1][1]);
                 $this->selection_end = $this->selection_start + strlen($matches[1][0]);
             }
         }
